Basic Frameworks
Town Script containing its many integers
Character and NPC integration nearing completion.
Node Created for Interactable.
Attach interactable to any node that is to be interacted with by the player.

Basic time of day system scripted in Town. Call the function to advance time, and it will emit a signal for what time of day it is