Cas movement init (#1)
* Basic Frameworks

Town Script containing its many integers
Character and NPC integration nearing completion.
Node Created for Interactable.
Attach interactable to any node that is to be interacted with by the player.

Basic time of day system scripted in Town. Call the function to advance time, and it will emit a signal for what time of day it is

* Cas initial player script and assets

* Player day/night switching base + moving between areas

---------

Co-authored-by: msernoskie <mark.sernoskie@gmail.com>