UI (#2)
* Basic Frameworks

Town Script containing its many integers
Character and NPC integration nearing completion.
Node Created for Interactable.
Attach interactable to any node that is to be interacted with by the player.

Basic time of day system scripted in Town. Call the function to advance time, and it will emit a signal for what time of day it is

* Exports

Added Exports to Town, Player, NPC
- Changing the alignment enum in Town will have it check all children for nodes that can have the alignment updated, and will trigger the changes.

* need to fix duplicate duplicate issues and controlling input away from rest of scenes issue

* lol

* several more ui wrapper components

* a

---------

Co-authored-by: msernoskie <mark.sernoskie@gmail.com>